# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MORIZO API — городские квесты
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quests:
    get:
      operationId: listQuests
      tags: [quests]
      summary: List published quests with filters
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: city
          schema: { type: string }
        - in: query
          name: difficultyMin
          schema: { type: integer, minimum: 1, maximum: 5 }
        - in: query
          name: difficultyMax
          schema: { type: integer, minimum: 1, maximum: 5 }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: nearLat
          schema: { type: number }
        - in: query
          name: nearLng
          schema: { type: number }
        - in: query
          name: radiusKm
          schema: { type: number }
        - in: query
          name: sort
          schema: { type: string, enum: [recent, popular, rating] }
        - in: query
          name: bestTravelMode
          schema: { type: string, enum: [foot, transport, public_transport, dirt_road, off_road] }
      responses:
        "200":
          description: Paginated quests
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuestListResponse" }
    post:
      operationId: createQuest
      tags: [quests]
      summary: Create quest (draft)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateQuestBody" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quest" }
...
  /quests/{id}:
    get:
      operationId: getQuest
      tags: [quests]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Quest detail (with checkpoints + author)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuestDetail" }
    patch:
      operationId: updateQuest
      tags: [quests]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateQuestBody" }
      responses:
        "200":
          description: Updated quest
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quest" }
  /quests/{id}/submit:
    post:
      operationId: submitQuest
      tags: [quests]
      summary: Submit draft for moderation
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated quest
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quest" }
...
  /quests/{id}/archive:
    post:
      operationId: archiveQuest
      tags: [quests]
      summary: Archive published quest
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated quest
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quest" }

  # ---------- CHECKPOINTS ----------
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      responses:
        "200":
          description: All public teams
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Team" }
    post:
      operationId: createTeam
      tags: [teams]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTeamBody" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Team" }
...
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Team detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeamDetail" }
  /teams/join:
    post:
      operationId: joinTeam
      tags: [teams]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JoinTeamBody" }
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Team" }

  # ---------- SESSIONS (passing) ----------
...
components:
  schemas:
...
    Quest:
      type: object
      required: [id, title, description, city, district, difficulty, durationMin, status, authorId, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        city: { type: string }
        district: { type: string }
        coverUrl: { type: string }
        difficulty: { type: integer, minimum: 1, maximum: 5 }
        durationMin: { type: integer }
        rules: { type: string }
        status:
          type: string
          enum: [draft, moderation, published, archived, rejected]
        rejectionReason: { type: string }
        authorId: { type: integer }
        authorNickname: { type: string }
        checkpointCount: { type: integer }
        completionCount: { type: integer }
        avgRating: { type: number }
        createdAt: { type: string, format: date-time }
        publishedAt: { type: string, format: date-time }
        bestTravelMode:
          type: string
          enum: [foot, transport, public_transport, dirt_road, off_road]
    QuestDetail:
      type: object
      required: [quest, checkpoints]
      properties:
        quest: { $ref: "#/components/schemas/Quest" }
        checkpoints:
          type: array
          items: { $ref: "#/components/schemas/Checkpoint" }
    QuestListResponse:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Quest" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
...
    CreateQuestBody:
      type: object
      required: [title, description, city, district, difficulty, durationMin]
      properties:
        title: { type: string, minLength: 5 }
        description: { type: string, minLength: 30 }
        city: { type: string }
        district: { type: string }
        coverUrl: { type: string }
        difficulty: { type: integer, minimum: 1, maximum: 5 }
        durationMin: { type: integer, minimum: 5 }
        rules: { type: string }
        bestTravelMode:
          type: string
          enum: [foot, transport, public_transport, dirt_road, off_road]
    UpdateQuestBody:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        city: { type: string }
        district: { type: string }
        coverUrl: { type: string }
        difficulty: { type: integer, minimum: 1, maximum: 5 }
        durationMin: { type: integer, minimum: 5 }
        rules: { type: string }
        bestTravelMode:
          type: string
          enum: [foot, transport, public_transport, dirt_road, off_road]
          nullable: true

    # ---------- CHECKPOINTS ----------
    Checkpoint:
      type: object
      required: [id, questId, orderIndex, name, task, taskType, lat, lng]
      properties:
        id: { type: integer }
        questId: { type: integer }
        orderIndex: { type: integer }
        name: { type: string }
        task: { type: string }
        taskType:
          type: string
          enum: [code_word, choice]
        codeAnswer: { type: string }
        choiceOptions:
          type: array
          items: { type: string }
        choiceAnswerIndex: { type: integer }
        hint: { type: string }
        rules: { type: string }
        lat: { type: number }
        lng: { type: number }
...
    Team:
      type: object
      required: [id, name, joinCode, captainId, memberCount, points, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        joinCode: { type: string }
        captainId: { type: integer }
        captainNickname: { type: string }
        memberCount: { type: integer }
        points: { type: integer }
        completedCount: { type: integer }
        createdAt: { type: string, format: date-time }
    TeamMember:
      type: object
      required: [id, nickname, role]
      properties:
        id: { type: integer }
        nickname: { type: string }
        role: { type: string, enum: [captain, member] }
    TeamDetail:
      type: object
      required: [team, members]
      properties:
        team: { $ref: "#/components/schemas/Team" }
        members:
          type: array
          items: { $ref: "#/components/schemas/TeamMember" }
    CreateTeamBody:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 2 }
        description: { type: string }
    JoinTeamBody:
      type: object
      required: [joinCode]
      properties:
        joinCode: { type: string }

    # ---------- SESSIONS ----------
...